reggie 点餐 前期项目完整版

+++ regiee_take_out_class/src/main/java/com/itheima/reggie/entity/OrderDetail.java
@@ -39,7 +39,7 @@ public class OrderDetail implements Serializable {
     private Integer number;
 
     //金额
-    private BigDecimal amount;
+    private Integer amount;
 
     //图片
     private String image;
